+++ src/logic/get-stop-places-with-lines/GetStopPlacesWithLines.graphql
@@ -1,0 +1,32 @@
+query getStopPlaces($ids: [String]!) @api(name: journey_planner_v3) {
+    stopPlaces(ids: $ids) {
+        id
+        name
+        description
+        latitude
+        longitude
+        transportMode
+        transportSubmode
+        estimatedCalls(
+            numberOfDeparturesPerLineAndDestinationDisplay: 1
+            timeRange: 604800
+            numberOfDepartures: 200
+            arrivalDeparture: departures
+        ) {
+            destinationDisplay {
+                frontText
+            }
+            serviceJourney {
+                line {
+                    id
+                    transportMode
+                    transportSubmode
+                    publicCode
+                }
+                pointsOnLink {
+                    points
+                }
+            }
+        }
+    }
+}
