+++ src/logic/useStopPlacesWithLines.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { Line, StopPlaceWithLines } from '../types'
 import { createAbortController, unique } from '../utils'
 import { useSettings } from '../settings/SettingsProvider'
-import { getStopPlacesWithLines } from './getStopPlacesWithLines'
+import { getStopPlacesWithLines } from './get-stop-places-with-lines/getStopPlacesWithLines'
 import { useStopPlacesWithDepartures } from './use-stop-places-with-departures/useStopPlacesWithDepartures'
 
 interface Return {
